replaced all LocalDateTime for LocalDate / started implementing the fee calculation strategy

+++ src/main/java/com/luisilva/transactionschedulerapp/exceptions/NoContentAtTheDatabaseException.java
@@ -10,7 +10,7 @@ public class NoContentAtTheDatabaseException extends RuntimeException {
         super("There is no content at the database of type: " + aClass.getName());
     }
 
-    public NoContentAtTheDatabaseException(Class<?> aClass, int id) {
+    public NoContentAtTheDatabaseException(Class<?> aClass, Long id) {
         super("There is no content at the database of type: " + aClass.getName() + " with client account id: " + id);
     }
 }
